[Feat.] Export refactoring
