+++ package.json
@@ -15,6 +15,7 @@
     "js-yaml": "^4.1.0",
     "jspdf": "^3.0.1",
     "jspdf-autotable": "^5.0.2",
+    "marked": "^15.0.7",
     "next": "15.2.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
